Model CURD

+++ src/main/java/com/bo/ModelBO.java
@@ -4,8 +4,8 @@ import lombok.Data;
 
 @Data
 public class ModelBO {
-    public Long modelId;
-    private Long manId;
+    public Integer modelId;
+    private Integer manId;
     private String modelName;
     private String year;
     private Integer seatNum;
