+++ src/main/java/com/vo/ModelVO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ModelVO {
-    public Long modelId;
+    public Integer modelId;
     private Long manId;
     private String modelName;
     private String year;
